Fix upstream menu to use the remote branch name
Previously the upstream options menu was using the local branch name.
This is fine as long as the local branch name matches the remote branch
name. However, if the local branch name and the remote branch name
differ, trying to use the reset or rebase feature will fail.

To fix the issue, ShortUpstreamRefName() is used when resolving the
remote branch name.