Add configurable behaviour for branch tracking
The current behaviour when creating a new branch off of a remote branch
is to always track the branch it was created from.

For example, if a branch 'my_branch' is created off of the remote branch
'fix-crash-13', then 'my_branch' will be tracking the remote
'fix-crash-13' branch.

It is common practice to have both the local and remote branches named
the same when the local is tracking the remote one. Therefore, it is
reasonable to expect that 'my_branch' should not track the remote
'fix-crash-13' branch.

The config 'Git.AutomaticTrackingWhenCreatingNewBranchFromRemoteBranch'
allows users to decide what will happen when creating a new branch off
of a remote one. The possible values and their behaviours are:

 - 'always' (default) => local branches will ALWAYS track the remote
   branches they were created from.

 - 'never' => local branches will NEVER track the remote branches they
   were created from.

 - 'whenBranchNamesMatch' => local branches will track the remote
   branches they were created from ONLY IF THEIR NAMES MATCH.